refactor bookmark update to implement updatable trait

+++ app/Http/Actions/Tag/SyncBookmarkTagsAction.php
@@ -13,37 +13,37 @@ final class SyncBookmarkTagsAction
 {
     public function __construct(private AttachOrCreateTagsAction $attachOrCreateTagsAction) {}
 
-    public function execute(Bookmark $bookmark, array $tags = []): void
+    public function execute(Bookmark $bookmark, array $tags = []): bool
     {
         $bookmarkTags = $bookmark->tags()->pluck('name')->toArray();
-
         $tagsToAttach = array_diff($tags, $bookmarkTags);
+        $tagsToDelete = array_diff($bookmarkTags, $tags);
+
+        $changed = false;
 
         if (count($tagsToAttach) >= 1) {
             $this->attachOrCreateTagsAction->execute(bookmark: $bookmark, tags: $tagsToAttach);
+            $changed = true;
         }
 
-        $this->deatachTags(
-            bookmark: $bookmark,
-            bookmarkTags: $bookmarkTags,
-            tags: $tags
-        );
+        if (count($tagsToDelete) >= 1) {
+            $this->detachTags(bookmark: $bookmark, tagsToDelete: $tagsToDelete);
+            $changed = true;
+        }
+
+        return $changed;
     }
 
-    private function deatachTags(
-        Bookmark $bookmark,
-        array $bookmarkTags,
-        array $tags
-    ): void {
-        $tagsToDelete = array_diff($bookmarkTags, $tags);
-
+    private function detachTags(Bookmark $bookmark, array $tagsToDelete): void
+    {
         DB::transaction(function () use ($bookmark, $tagsToDelete) {
             foreach ($tagsToDelete as $tag) {
                 $tagModel = Tag::whereUserId(Auth::id())->where('name', $tag)->first();
-                $bookmark->tags()->detach($tagModel);
-
-                if ($tagModel->bookmarks()->count() === 0) {
-                    $tagModel->delete();
+                if ($tagModel) {
+                    $bookmark->tags()->detach($tagModel);
+                    if ($tagModel->bookmarks()->count() === 0) {
+                        $tagModel->delete();
+                    }
                 }
             }
         });
